+++ src/patch/dll_collision_viewer.zig
@@ -6,7 +6,6 @@ const w32wm = w32.ui.windows_and_messaging;
 const XINPUT_GAMEPAD_BUTTON_INDEX = @import("core/Input.zig").XINPUT_GAMEPAD_BUTTON_INDEX;
 const VIRTUAL_KEY = w32.ui.input.keyboard_and_mouse.VIRTUAL_KEY;
 
-const GlobalSt = @import("appinfo.zig").GLOBAL_STATE;
 const GlobalFn = @import("appinfo.zig").GLOBAL_FUNCTION;
 const COMPATIBILITY_VERSION = @import("appinfo.zig").COMPATIBILITY_VERSION;
 const VERSION_STR = @import("appinfo.zig").VERSION_STR;
@@ -357,7 +356,7 @@ export fn PluginCompatibilityVersion() callconv(.C) u32 {
     return COMPATIBILITY_VERSION;
 }
 
-export fn OnInit(_: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
+export fn OnInit(gf: *GlobalFn) callconv(.C) void {
     AnnodueSettings.settingsInit(gf);
 
     init_collision_viewer(&state);
@@ -365,21 +364,21 @@ export fn OnInit(_: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
     QuickRaceMenu.gf = gf;
 }
 
-export fn OnInitLate(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
+export fn OnInitLate(_: *GlobalFn) callconv(.C) void {}
 
-export fn OnDeinit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
+export fn OnDeinit(_: *GlobalFn) callconv(.C) void {
     QuickRaceMenu.close();
     deinit_collision_viewer();
 }
 
 // HOOKS
 
-export fn InputUpdateB(_: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
+export fn InputUpdateB(gf: *GlobalFn) callconv(.C) void {
     input_enable.update(gf);
     input_pause.update(gf);
     QuickRaceMenu.update_input();
 }
 
-export fn EarlyEngineUpdateB(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
+export fn EarlyEngineUpdateB(_: *GlobalFn) callconv(.C) void {
     QuickRaceMenu.update();
 }