+++ src/patch/dll_cosmetic.zig
@@ -2,7 +2,6 @@ const Self = @This();
 
 const std = @import("std");
 
-const GlobalSt = @import("appinfo.zig").GLOBAL_STATE;
 const GlobalFn = @import("appinfo.zig").GLOBAL_FUNCTION;
 const COMPATIBILITY_VERSION = @import("appinfo.zig").COMPATIBILITY_VERSION;
 const VERSION_STR = @import("appinfo.zig").VERSION_STR;
@@ -379,7 +378,7 @@ export fn PluginCompatibilityVersion() callconv(.C) u32 {
     return COMPATIBILITY_VERSION;
 }
 
-export fn OnInit(_: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
+export fn OnInit(gf: *GlobalFn) callconv(.C) void {
     CosmeticState.settingsInit(gf);
 
     // TODO: convert to use global allocator once it is part of the GlobalFn interface;
@@ -409,9 +408,9 @@ export fn OnInit(_: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
     //gs.patch_offset = off;
 }
 
-export fn OnInitLate(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
+export fn OnInitLate(_: *GlobalFn) callconv(.C) void {}
 
-export fn OnDeinit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
+export fn OnDeinit(_: *GlobalFn) callconv(.C) void {
     crot.PatchRgbArgs(0x460E5D, 0xFFFFFF); // in-race hud UI numbers
     crot.PatchRgbArgs(0x460FB1, 0xFFFFFF);
     crot.PatchRgbArgs(0x461045, 0xFFFFFF);
@@ -423,7 +422,7 @@ export fn OnDeinit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
 
 // HOOKS
 
-export fn TextRenderB(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
+export fn TextRenderB(_: *GlobalFn) callconv(.C) void {
     if (CosmeticState.s_rb_enable) {
         CosmeticState.PatchHudColRotate(
             CosmeticState.s_rb_value_enable,